refactor: 클래스 모듈 파일 분리
