+++ src/modules/LottoGenerator.js
@@ -1,5 +1,5 @@
 const MissionUtils = require('@woowacourse/mission-utils');
-const LottoCost = require('./LottoCost');
+const LottoCost = require('../validation/LottoCost');
 
 class LottoGenerator {
   static createLottoNums() {
