+++ src/utils/Input.js
@@ -1,13 +1,13 @@
 const { Console } = require('@woowacourse/mission-utils');
 const GAME_MESSAGE = require('../constant/gameMessage');
-const LottoGenerator = require('../LottoGenerator');
-const Lotto = require('../Lotto');
-const WinNums = require('../WinNums');
-const BonusNum = require('../BonusNum');
-const LottoMatching = require('../LottoMatching');
+const LottoGenerator = require('../modules/LottoGenerator');
+const Lotto = require('../validation/Lotto');
+const WinNums = require('../validation/WinNums');
+const BonusNum = require('../validation/BonusNum');
+const LottoMatching = require('../modules/LottoMatching');
 const View = require('./View');
 
-class Input {  
+class Input {
   static lottoCost() {
     Console.readLine(GAME_MESSAGE.LOTTO_COST_INPUT_MESSAGE, (lottoCost) => {
       const [lottoCount, lottoArr] = LottoGenerator.publishLotto(lottoCost);
@@ -29,8 +29,15 @@ class Input {
   static bonusNum(lottoArr, winNums, lottoCost) {
     Console.readLine(GAME_MESSAGE.BONUS_NUM_INPUT_MESSAGE, (bonusNum) => {
       new BonusNum(winNums, bonusNum);
-      const lottoResultObj = LottoMatching.getResultObj(lottoArr, winNums, bonusNum);
-      const rateOfReturn = LottoMatching.getRateOfReturn(lottoResultObj, lottoCost);
+      const lottoResultObj = LottoMatching.getResultObj(
+        lottoArr,
+        winNums,
+        bonusNum
+      );
+      const rateOfReturn = LottoMatching.getRateOfReturn(
+        lottoResultObj,
+        lottoCost
+      );
       View.winStatistics(lottoResultObj, rateOfReturn);
     });
   }
